requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gspread
google-auth
python-dotenv
schedule==1.2.1
gunicorn
pytz
